ci(bench): split check-gate environment resolution into its own module
